+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # HEAD
-
+- Added analytical gaussian alm provider which can include ellipticity. 
 - Added new tests [[#1](https://github.com/teob97/grasp2alm/pull/1)]
 - Added a complete doumentation [[#4](https://github.com/teob97/grasp2alm/pull/4)]